# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/register": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/create-professor": {
      "post": {
        "description": "Permite que um super administrador crie um novo professor sem senha. Um email será enviado com credenciais temporárias.",
        "operationId": "UserController_createProfessor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfessorByAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Professor criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "description": "Dados do usuário criado"
                    },
                    "temporaryPassword": {
                      "type": "string",
                      "description": "Senha temporária gerada (para casos onde o email falha)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou email/matrícula já existem"
          },
          "403": {
            "description": "Usuário não tem permissões de super administrador"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar professor por super administrador",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/delete/{id}": {
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/toggle-activation/{id}": {
      "patch": {
        "operationId": "UserController_toggleUserActivation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activate",
            "required": true,
            "in": "query",
            "description": "Activate or deactivate the user.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/users/{id}/approve": {
      "patch": {
        "operationId": "UserController_approveTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/approve-presenter": {
      "patch": {
        "operationId": "UserController_approvePresenter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UserController_getUsers",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by user status (e.g., Active, Inactive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive search on name/email (server-side).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profiles",
            "required": false,
            "in": "query",
            "description": "Filter by profiles (e.g., Professor, Listener). Accepts multiple values.",
            "schema": {}
          },
          {
            "name": "roles",
            "required": false,
            "in": "query",
            "description": "Filter by user levels (e.g., Admin, Default). Accepts multiple values.",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/advisors": {
      "get": {
        "operationId": "UserController_getAdvisors",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/admins": {
      "get": {
        "operationId": "UserController_getAdmins",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/confirm-email": {
      "post": {
        "operationId": "UserController_confirmEmail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/edit/{email}": {
      "patch": {
        "operationId": "UserController_editUserBySuperAdmin",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "WEPGCOMP API",
    "description": "Portal WEPGCOMP - API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "CreateProfessorByAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome completo do professor",
            "example": "Dr. João Silva",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "description": "Email do professor (deve ser @ufba.br)",
            "example": "joao.silva@ufba.br",
            "maxLength": 255
          },
          "registrationNumber": {
            "type": "string",
            "description": "Número de matrícula SIAPE do professor (até 13 dígitos)",
            "example": "1234567890123",
            "pattern": "^\\d{1,13}$",
            "maxLength": 13
          },
          "registrationNumberType": {
            "type": "string",
            "description": "Tipo do número de registro",
            "enum": [
              "CPF",
              "MATRICULA"
            ],
            "example": "MATRICULA"
          },
          "photoFilePath": {
            "type": "string",
            "description": "Caminho para foto do perfil (opcional)",
            "example": "/uploads/profile/photo.jpg",
            "maxLength": 255
          }
        },
        "required": [
          "name",
          "email",
          "registrationNumber"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}